move everything within stack package
This removes the `shell` package moving the logic alongside
`ShellInit`, removes the `Shell` type and reduce template code
to only POSIX compliant and fish shell.
