--- internal/stack/shellinit/shell/templates.go
@@ -1,19 +1,0 @@
-package shell
-
-const BashInitCode = ShInitCode
-
-const FishInitCode = `set -x %s %s;
-set -x %s %s;
-set -x %s %s;
-set -x %s %s;
-set -x %s %s;
-`
-
-const ShInitCode = `export %s=%s
-export %s=%s
-export %s=%s
-export %s=%s
-export %s=%s
-`
-
-const ZshInitCode = ShInitCode
